Move all config classes to single file `config.py`
Sets up a single config file with the classes `Config`,
`ProductionConfig`, `DevConfig` and `TestingConfig`. The correct config
class will be chosen at runtime based on the FLASK_ENV environment
variable.

This is in an attempt to fix the errors in the Travis builds when it
attempted to load the config file from an on-system json file.